+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## main
 
+## 3.1.3
+
+- Fix: now using the reserved space for the OPT RR records during the encoding of the message (#74)
+
+## 3.1.2
+
+- Add NXNAME type to terms types.
+
+## 3.1.1
+
+- Fix EDNS0 compliance for truncated records and unsupported versions
+
+## 3.1.0
+
+- Fix EDNS0 compliance for truncated records and unsupported versions
+
 ## 3.0.5
 
 - Fix `max_size` in encode_message opts